Fix header parsing bugs and panics from code review
A7800: the A78 header has no checksum field; the parser was summing
the title bytes as a "stored checksum" (Desert Falcon reported
6F6E2020, which is "on  " from the title padding). Drop the phantom
StoredChecksum impl, read the cartridge type as the 16-bit big-endian
bitfield at 0x35 (was a single byte at 0x18), and trim space padding
from titles.

NES: decode the mapper using only the header fields valid for the
detected format. Byte 8 contributes mapper bits 8-11 only in NES 2.0;
iNES 1.0 files with a nonzero PRG-RAM byte gained phantom high bits.
Archaic iNES trusts only the byte-6 nibble, since byte 7 often holds
junk like "DiskDude!". Also raise the UxROM PRG limit to the 4MB an
8-bit bank register addresses; 256K only applies to the original
discrete UNROM/UOROM boards and falsely flagged valid 512K homebrew.

GB: the RAM size header byte is a lookup table, not a shift; codes
0/2/3 mean 0/8K/32K, not the 1K/4K/8K the old shift produced. ROM
size keeps the shift for codes 0x00-0x08 and gains the unofficial
0x52-0x54 entries; unknown codes now print as unknown rather than a
garbage size.

SMS: 48KB ROMs (size code 0xD) checksum the third bank at
0x8000-0xBFFF in addition to the first 32KB; they were lumped in with
the sizes that only sum the first banks.

N64: print the boot CRC pair as full 8-digit words (leading zeros
were dropped), require the full 1MiB CRC window up front instead of
0x1000 bytes, and return Cow from byte-order correction so
big-endian ROMs skip a full copy.

FDS: identify Quick Disk containers by the authoritative block-2
marker at 0x3A before falling back to .FDS; a QD whose block-1 CRC
low byte is 0x02 was misread as packed .FDS. Guard Display against a
default-constructed value with no sides, and widen disk numbers
before adding 1.

Registry: registrations gain an optional max_size so auto-detection
skips systems whose largest ROM is far smaller than the input
(game.com 4MB, ColecoVision 1MB), avoiding signature scans over large
disc images. Explicit --system selection is unaffected. Entries now
use a const builder instead of repeating every field.

Also: Genesis ROM size widens before +1 so a 0xFFFFFFFF end address
cannot overflow, GameCube/Wii disc numbers widen before +1, dat
output takes its <version> from CARGO_PKG_VERSION instead of a
hardcoded 0.1.0, and dead code (verify_sega_checksum, unused cue
sector size) is removed.

Verified with a local before/after snapshot over 681 ROMs: 40 outputs
changed, each traced to one of the fixes above, and no file changed
which system it detects as. fmt, clippy -D warnings, and all 51
tests are green.

+++ src/systems/disc/nintendo_disc.rs
@@ -145,7 +145,7 @@ impl std::fmt::Display for NintendoDiscHeader {
         writeln!(f, "Game ID: {}", self.game_id)?;
         writeln!(f, "Region: {}", decode_region(self.region_code))?;
         writeln!(f, "Version: 1.{:02}", self.version)?;
-        writeln!(f, "Disc Number: {}", self.disc_number + 1)?;
+        writeln!(f, "Disc Number: {}", u16::from(self.disc_number) + 1)?;
         if self.audio_streaming {
             writeln!(f, "Audio Streaming: Enabled")?;
         }
